web-frontend/test/requirements.txt
Click
fastapi
uvicorn[standard]==0.17.*
gunicorn
itsdangerous
Jinja2
MarkupSafe
SQLAlchemy
Werkzeug
llama_index
torch
transformers
panda
numpy
langchain
python-dotenv
python-multipart
httpx
openai